Move ActionButton.kt to the right package
